+++ lib/types/message/meta/meta.dart
@@ -1,6 +1,7 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'meta.freezed.dart';
+part 'meta.g.dart';
 
 class MessageMetaData {
   List<MetaDataEntry> favorites;
@@ -18,4 +19,7 @@ class MetaDataEntry with _$MetaDataEntry {
       {required String id,
       required String name,
       String? classOrInitials}) = _MetaDataEntry;
+
+  factory MetaDataEntry.fromJson(Map<String, Object?> json) =>
+      _$MetaDataEntryFromJson(json);
 }