+++ lib/types/message/meta/meta.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'meta.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_MetaDataEntry _$$_MetaDataEntryFromJson(Map<String, dynamic> json) =>
+    _$_MetaDataEntry(
+      id: json['id'] as String,
+      name: json['name'] as String,
+      classOrInitials: json['classOrInitials'] as String?,
+    );
+
+Map<String, dynamic> _$$_MetaDataEntryToJson(_$_MetaDataEntry instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'name': instance.name,
+      'classOrInitials': instance.classOrInitials,
+    };